Обновление лабораторной #1
Обновление лабораторной, чтобы была возможность запустить через Maven в соостветии с быстрым стартом
