--- lab01/source/HelloWorldApp.java
@@ -1,7 +1,0 @@
-class HelloWorldApp
-{
-	public static void main(String[] args)
-	{
-		System.out.println("Hello World!");
-	}
-}
